+++ library/core/src/main/java/com/google/android/exoplayer2/trackselection/BufferSizeAdaptationBuilder.java
@@ -29,7 +29,6 @@ import com.google.android.exoplayer2.upstream.BandwidthMeter;
 import com.google.android.exoplayer2.upstream.DefaultAllocator;
 import com.google.android.exoplayer2.util.Assertions;
 import com.google.android.exoplayer2.util.Clock;
-import com.google.android.exoplayer2.util.PriorityTaskManager;
 import java.util.List;
 import org.checkerframework.checker.nullness.compatqual.NullableType;
 
@@ -112,7 +111,6 @@ public final class BufferSizeAdaptationBuilder {
   private int hysteresisBufferMs;
   private float startUpBandwidthFraction;
   private int startUpMinBufferForQualityIncreaseMs;
-  @Nullable private PriorityTaskManager priorityTaskManager;
   private DynamicFormatFilter dynamicFormatFilter;
   private boolean buildCalled;
 
@@ -220,20 +218,6 @@ public final class BufferSizeAdaptationBuilder {
   }
 
   /**
-   * Sets the {@link PriorityTaskManager} to use.
-   *
-   * @param priorityTaskManager The {@link PriorityTaskManager}.
-   * @return This builder, for convenience.
-   * @throws IllegalStateException If {@link #buildPlayerComponents()} has already been called.
-   */
-  public BufferSizeAdaptationBuilder setPriorityTaskManager(
-      PriorityTaskManager priorityTaskManager) {
-    Assertions.checkState(!buildCalled);
-    this.priorityTaskManager = priorityTaskManager;
-    return this;
-  }
-
-  /**
    * Sets the {@link DynamicFormatFilter} to use when updating the selected track.
    *
    * @param dynamicFormatFilter The {@link DynamicFormatFilter}.
@@ -266,9 +250,6 @@ public final class BufferSizeAdaptationBuilder {
                 maxBufferMs,
                 bufferForPlaybackMs,
                 bufferForPlaybackAfterRebufferMs);
-    if (priorityTaskManager != null) {
-      loadControlBuilder.setPriorityTaskManager(priorityTaskManager);
-    }
     if (allocator != null) {
       loadControlBuilder.setAllocator(allocator);
     }